Refactored qdebugstream to QDebugStream to follow our CamelCase conventions #195
