+++ gui/src/QDebugStream.h
